requirements.txt
Django==4.2.4
PyGithub==2.1.1
PyGithub==2.1.1
